Migrate currentAttemptUpdate code to shared code
